feat: calendar 데이터 페칭을 커스텀훅으로 분리하고 로딩상태 추가 [#19]
