+++ src/components/calendar/calendar-contents.tsx
@@ -3,13 +3,17 @@ import CalendarContent from "../post/calendar-content";
 
 export default function CalendarContents() {
   const {
-    posts,
+    posts: { data, isLoading },
     calendar: { selectedDate, currentDate },
   } = useCalendarContext();
 
   const [year, month, date] = selectedDate.date.split("-");
 
-  if (posts.length === 0) {
+  if (isLoading) {
+    return <div className="shadow-xl rounded-xl p-4"></div>;
+  }
+
+  if (!isLoading && data.length === 0) {
     return (
       <div className="w-full text-center text-lg font-semibold text-underline mt-10">{`${currentDate.month}달 인증이 존재하지 않습니다.`}</div>
     );
@@ -19,7 +23,7 @@ export default function CalendarContents() {
       {selectedDate.date ? (
         <>
           <div className="text-center text-lg font-semibold">{`${year}년 ${month}월 ${date}일 인증`}</div>
-          {posts
+          {data
             .find((post) => post.date === selectedDate.date)
             ?.posts.map((post) => (
               <CalendarContent key={post.id} post={post} />
@@ -28,7 +32,7 @@ export default function CalendarContents() {
       ) : (
         <>
           <div className="text-center text-lg font-semibold">{`${currentDate.month}월 인증`}</div>
-          {posts
+          {data
             .flatMap((post) => post.posts)
             .map((post) => (
               <CalendarContent key={post.id} post={post} />
