+++ src/components/calendar/index.tsx
@@ -1,29 +1,22 @@
 "use client";
 
-import { ReactNode, useEffect, useState } from "react";
+import { ReactNode } from "react";
+
 import { CalendarContext } from "./useCalendarContext";
 import CalendarHeader from "./calendar-header";
 import CalendarBody from "./calendar-body";
 import SelectedDate from "./seleted-date";
 import useCalendar from "@/hooks/useCalendar";
 import CalendarContents from "./calendar-contents";
-import { getWrittenPostByYearnAndMonth } from "@/app/(tab)/calendar/actions";
-import { GroupPostsByDateType } from "@/lib/utils";
+import useGetCalendarPosts from "@/hooks/useGetCalendarPosts";
 
 const Calendar = ({ children }: { children: ReactNode }) => {
   const calendar = useCalendar();
   const { year, month } = calendar.currentDate;
-  const [calendarPosts, setCalendarPosts] = useState<GroupPostsByDateType>([]);
-  useEffect(() => {
-    const fetchingPostsByCalendar = async () => {
-      const posts = await getWrittenPostByYearnAndMonth(Number(year), Number(month));
-      setCalendarPosts(posts);
-    };
-    fetchingPostsByCalendar();
-  }, [year, month]);
+  const { calendarPosts, isLoading } = useGetCalendarPosts(Number(year), Number(month));
 
   return (
-    <CalendarContext.Provider value={{ calendar, posts: calendarPosts }}>
+    <CalendarContext.Provider value={{ calendar, posts: { data: calendarPosts, isLoading } }}>
       <div className="flex flex-col gap-2 w-full">{children}</div>
     </CalendarContext.Provider>
   );
